(+,-) import h5py in plotting script, version from requirements
Added an import of the h5py package to the winnet_class plotting script
which was missing to plot the r_process_Obergaulinger_example. Removed
the python package versions from the requirements. Newer Anaconda
versions have problems with the old ones. Still the tested versions of
python packages are listed in requirements_tested.