(+,-,*) Documentation, typos, time step for implicit euler (#8)
* (+) restriction to gear solver's timestep

The gear solver was in theory able to increase the time step without
restriction. Now it is restricted by a new parameter (gear_timestep_max)
which is set to 10 by default.

* (*) tolerance for eq_emission test

WIth the timestep changes the finab of this test changed slightly with
gfortran. Therefore the tolerance increased from 5e-4 to 1e-3

* (*) typo  in error message

The possible solver flag was not correct. It was written that solver 1
or 2 is possible, but actually it is 0 and 1.

* (+) Short explanation for Windows installation

Added a short documentation on how to run WinNet on Windows. I believe
the easiest way is to use the Windows Subsystem for Linux (WSL). I
tested it and it worked right away without any problem.

* (+) documentation in benam_class

This documentation was forgotten before...

* (-) typo in output message

* (-) Typo in documentation

* (-) Bug in timestepping

The estimate for the timestep in the implicit euler was only restricted
for creation, not for destruction (there was an absolute missing).
Introduced this absolute and modified the test cases accordingly.

* (+,-) import h5py in plotting script, version from requirements

Added an import of the h5py package to the winnet_class plotting script
which was missing to plot the r_process_Obergaulinger_example. Removed
the python package versions from the requirements. Newer Anaconda
versions have problems with the old ones. Still the tested versions of
python packages are listed in requirements_tested.