(*) Rewriting of top_engen_every parameter
The top_engen_every parameter was only working when nuclear heating was
turned on (and even then it was called several times instead of one
time). Moved everything to analysis.f90 and also implemented that not
only the changing isotopes but also the top contributing reactions are
outputted. All these changes also have been described in the
documentation of the parameters.