(+) automatic docker file via gh packages
Added workflow to automatically create and deploy a docker image with a
working WinNet environment. This environment contains the intel fortran
compiler and the hdf5 library compiled with ifx. The workflow is
triggered whenever a new release is published and the package will be
visible under "packages" on the github page. A instruction on how to use
this is added to the documentation.