Automatic package creation on release. (#10)
* (+) automatic docker file via gh packages

Added workflow to automatically create and deploy a docker image with a
working WinNet environment. This environment contains the intel fortran
compiler and the hdf5 library compiled with ifx. The workflow is
triggered whenever a new release is published and the package will be
visible under "packages" on the github page. An instruction on how to use
this is added to the documentation.

* (-) bug in makerun.py.

The folders "snaps/" and "flow/" were not created by the makerun when only the
custom_snapshot parameter was present and not the "snaps_every" or
"flow_every". This resulted in an error when trying to run winnet as 
winnet was unable to save the snapshots.