(+) Major improvement in fission
Fission rates got split into different files for different types of
fission. This removes the necessity of having hardcoded flags in the
code. Furthermore, beta-delayed fission can now be given in terms of
probability from the total beta-decay rate (as given mostly in
literature). Additionally, added the possibility to have beta-delayed
neutron emission fission. The different fission types can be given in
different formats that are specified with new parameters. Unfortunately,
this commit breaks compatibility to the older versions as the paramter
fission_rates got removed. The example cases still have to be adapted
accordingly...