(*) Rehaul of fission treatment into different channels (#11)
* (*) Rewriting of top_engen_every parameter

The top_engen_every parameter was only working when nuclear heating was
turned on (and even then it was called several times instead of one
time). Moved everything to analysis.f90 and also implemented that not
only the changing isotopes but also the top contributing reactions are
outputted. All these changes also have been described in the
documentation of the parameters.

* (*) Small format change in nuclear energy output

Appearantly it could happen that a number like 1....-101 (without E) for
the format were printed when number smaller than 1e-99 occured. This makes
it difficult to read the number later. Introduced a third digit for the
exponent.

* (-) bug in tabulated rates

The amount of counted tabulated rates and actual read tabulated rates
could mismatch resulting in an later not understandable error.
Fixed it to count tabulated rates correctly....

* (+) Major improvement in fission

Fission rates got split into different files for different types of
fission. This removes the necessity of having hardcoded flags in the
code. Furthermore, beta-delayed fission can now be given in terms of
probability from the total beta-decay rate (as given mostly in
literature). Additionally, added the possibility to have beta-delayed
neutron emission fission. The different fission types can be given in
different formats that are specified with new parameters. Unfortunately,
this commit breaks compatibility to the older versions as the paramter
fission_rates got removed. The example cases still have to be adapted
accordingly...

* (*,+) Fission module, new fissflag

Restructered fission more, kicked out rrate_fiss as it was not used.
Made fission fragment subroutines easier to read. Added more comments.
Added a new fissflag (4) to completely customize your fragment distribution,
this comes with a bunch of new parameters (sfission_file,
fission_frag_spontaneous,...).

* (*) Parameter files for example cases and documentation

Changed the example cases to work with the new parameters for splitted
fission treatment. Added beta-delayed fission rate file of Mumpower
et al. 2022. Changed the documentation and description accordingly.

* (+) variable tabulated temperature grid

* (*) made tabulated_rate non-private for compatability with rate_variation

* (+) documentation, monotonicity check, test case

Added a check for monotonicity in the tabulated temperature grid. Made
the grid a bit more flexible in the sense that it does not have to be
separated by exactly one space. Added test case.

* (*) Used wrong error code

* (-) bugs from merging

The fission needed access to the tabulated rates and their treatment had
changed. Made changes to make everything work again.

* (+) error message for fission probabilities larger than 1.

* (*) default value of beta-delayed fission format

Now the default value is 2: probability format

* (*) Fixed rosswog fission example

The fissflag was accidentally set wrong when implementing the fission
treatment.

* (-) bug when compiling with gfortran

One statement in the legacy abla subroutines had an "inout" flag.
Changed it to intent(in) since it is not changed. This has no impact in
any case since the routine is not used....

---------

Co-authored-by: Max Jacobi <mjacobi@theorie.ikp.physik.tu-darmstadt.de>