Heating convergence changes (#14)
* (+) features in makerun.py, conv. parameter for heating

Added environment variable in makerun.py to save the WinNet path for
default parameters. Added also the amount of OMP threads there which is
per default now set to 1. The heating got a new convergence control
parameter named heating_T9_tol and the temperature is brought to
convergence by limiting the shooting by 1/nr_count.

* (+) error message for wrong initial abundances

When the initial abundances deviate too much from sum of mass fractions
equal 1, now an error will be raised. Also had to change some test cases
as I was using the renormalization there.

* (*) tests only on one thread as well

Added the OMP_NUM_THREADS statement to the tests as well...

* (+) OMP threads and env variable also set for prepare mode

Also set the threads and environment variable for the winnet path to be
known when the makerun is run with --prepare.