Movie script (#15)
* (-) typos in docu

Removed typos in the documentation that Jan made me aware of.

* (*) Default number of threads=1

Set the default number of OMP threads to 1 in case OMP_NUM_THREADS was
not set. Before it was taking the maximum cores. Since PARDISO does not
parallize well, 1 is the better choice.

* (*) Ascii flow not outputted in zeroth iteration

The flow was producing an empty file in the zeroth iteration as there
was no flow yet. Removed this output.

* (+) movie script

Added a script to create a movie of a WinNet run. Therefore also changed
the Dockerfile that all dependencies are installed in docker as well.