Movie update (#16)
* (*) Tabulated rate reading

The format of reading the tabulated rates was not very general (f10.3).
Changed it to read(id,*). Also the tabulated temperature grid file could
get truncated for large grids. Used now a large character to read
(10000).

* (+) ascl identifier badge

Added a line in the Readme.md to add the badge and link to ASCL.

* (+) Citation file

Added citation file CITATION.cff for easier citing.

* (*) slight changes in the movie

Changed the requirements for python packages and made wreader a bit more
powerful (the new features are however not used).

* (*) Movie script and docu

Changed the movie script, now the flow arrows can change size.
Furthermore introduced the possibility of plot energy or tracked nuclei.
Additionally added a gif of one of these movies in the documentation
(mainpage and readme in the movie script folder).

* (*) Documentation in readme of movie script.

Added more detailed explanation of which parameters are necessary for
the movie.

* (-) typo

misstyped something in the docu.

* (+,-) Possibility to save frames in movie, annoying warning

Added a parameter for saving frames of the movie script. Removed a
warning (creating array temporary) when saving flows in hdf5 format.

* (*) Docu of movie script and turning of minor ticks as well

Changed the documentation of the movie script for a better explanation
of the --additional_plot option. Furthermore, also minor ticks are
turned off now in case someone chaged his matplotlibrc file.