(+) Script to summarize multiple runs into one large hdf5 file (#19)
* (+) Script to summarize many trajectories.

Added a script that is able to summarize a run that has been invoked
with the --many option. It will create an hdf5 file that contains
the enabled data on a common time grid (linearly interpolated).
Furthermore, custom_snapshots and final abundances are written into the
same file.

* (-,*) Bug, summary script

Removed bug in analysis.f90 that prevented some outputs from having an
output for the final time. Changed the summary script to also include
neutrino losses. Furthermore, reordered the relative imports for it. Its
not the best as it is now done but at least it is functioning.