(+) brents method for inverse interpolation
In rare occasions it could happen that the NR could not find the start
temperature (or at least not the last one and it ended in a wrong
minimum). To avoid that, I introduced brents method that respects
boundary problems. This should work more reliable than the NR for the
start temperature.