Dev (#28)
* (+) brents method for inverse interpolation

In rare occasions it could happen that the NR could not find the start
temperature (or at least not the last one and it ended in a wrong
minimum). To avoid that, I introduced brents method that respects
boundary problems. This should work more reliable than the NR for the
start temperature.

* (-) annoying warning in hdf5 save

When writing the finab to the hdf5 there was a warning that a function
got a array temporary. There was no problem with it, just the warning. I
removed the warning by writing the argument explicitely as variable and
not as Y*A.

* (-) bug in summary script

The mapping of the final abundances was buggy... Fixed that with a
drawback on performance...

* (+) performance of summary script

enhanced the mapping to improve the performance of the summary script.
Now it is a searchsorted and not a loop anymore.

* (-) bug when using mumpower fission file

There was a bug when using beta delayed fission with additional neutron
decay. Sometimes the neutrons where not taken into account properly
(when synthesizing something behind the dripline). Fixed that and made
the fragments bullet proof with an additional error.

* (*) summary_script. Take last index for run id.

Usually a run id is written last in a trajectory name. Before for the
example tra_ns2_144.dat would have given 2 as run id. Changed it that it
will give 144.