(*) Summary script (#30)
* (*) summary script

Now the tracked nuclei are named as in WinNet. E.g., protons are
'p' instead of 'h1', same for deuterium and tritium.

* (*) Summary script to correctly summarize custom snapshots

The custom snapshots were not reflecting a missing time (e.g., when the
trajectory started later than a time given in custom_snapshots) with a
nan value. Now it is shown correctly as nan in the hdf5 file.