(+) parameter nu_max_time to shut off neutrinos
This parameter can be useful for e.g., CCSN simulations because at the
moment winnet will just keep nu-luminosities and nu-temperatures
constant after the simulation has ended. If the tracer is not flying
away (i.e., not increasing the radius) this can lead to a very high
neutrino radiation over time. For CCSN this parameter should be maybe
set to 1d2, per default it is -1 and off.