(*,+) Movie and summary changes, new parameter to shut off neutrino reactions. (#32)
* (+) parameter nu_max_time to shut off neutrinos

This parameter can be useful for e.g., CCSN simulations because at the
moment winnet will just keep nu-luminosities and nu-temperatures
constant after the simulation has ended. If the tracer is not flying
away (i.e., not increasing the radius) this can lead to a very high
neutrino radiation over time. For CCSN this parameter should be maybe
set to 1d2, per default it is -1 and off.

* (*) docu

Changed the description of default parameter from nu_max_time.

* (+) WinNet movie got time slider.

Now, thanks to H. Rose, there is a timeslider option for the WinNet
movie. When creating a movie you can add --slider to add it and it makes
it easier to jump to specific frames.

* (+) Play button for the movie

Added a playbutton for the movie. Also spacebar is now playing and
stopping it.

* (+) functionality to interactive movie

Added switches to show timescales etc with buttons

* (*) More powerful interactive mode of movie

Added several things to the interactive mode of the movie. Further,
renamed the option to --interactive instead of --slider.

* (+) recursive option to summary script.

The summary script can not summarize a folder that contains folders of
runs. It will create one hdf5 file per subfolder.

* (+) features to interactive movie

Added data when clicked on the plot

* (-) bugfix in interactive movie

The r-process path was sometimes not shown properly when movie was
stopped.

* (*) made summary script to also summarize detailed engen

Before, if detailed energy generation was enabled, the summary script
was failing. Now it is able to summarize also this. Note that this can
be a huge amount of data.

* (+) options for interactive movie

Now it is possible to change the background and to toggle
flows/massfractions and also to change the range of mass fractions

* (+) description of interactive mode, poss. to turn off flows

Added describtion to the readme of the flow plot. Also added a
possibility to turn on/off flow arrows and mass fractions